Aggressive clean_up script

+++ clean_up.py
@@ -1,5 +1,6 @@
 ### Clears out the game image directories
 import os
+import shutil
 
 def delete_files():
     root_path = 'game'
@@ -10,14 +11,21 @@ def delete_files():
         working_path = os.path.join(root_path, folder_name)
 
         for item in os.listdir(working_path):
-            full_path = os.path.join(working_path,item)
+            full_path = os.path.join(working_path, item)
+
+            if os.path.basename(full_path) == 'EMPTY.md':
+                continue
 
             if os.path.isfile(full_path):
                 os.remove(full_path)
-                print(f'Deleted {full_path}')
-                i+=1
+                print(f'Deleted file {full_path}')
+                i += 1
+            elif os.path.isdir(full_path):
+                shutil.rmtree(full_path)
+                print(f'Deleted directory {full_path}')
+                i += 1
 
-    print(f'Deleted {i} item{bool(i-1)*"s"}')
+    print(f'Deleted {i} item{"s" if i != 1 else ""}')
 
 if __name__ == '__main__':
     delete_files()
